[lliu41] Company API + EF core  (#3)
* Feat: Implement add company service

* Feat: implement GET method

* Feat: Add profile entity to company entity

* Feat: Add employee to company entity

* Feat: implement delete API

* BugFix: include employee&profile data in company when retrieve

* Test: use inmemory Db for tests

* Test: check service create company

* Test: check service get company by id

* Test: check service get all companies

* Test: check service delete company by id

* Test: use random InMemory DB

* Refactor: use foreign key for delete

* Feat: use InMemoryDB for each method and each test class

+++ EFCoreRelationshipsPractice/Startup.cs
@@ -52,8 +52,15 @@ namespace EFCoreRelationshipsPractice
             {
                 using (var context = scope.ServiceProvider.GetService<CompanyDbContext>())
                 {
-                    context.Database.EnsureDeleted();
-                    context.Database.EnsureCreated();
+                    if (context.Database.ProviderName == "Microsoft.EntityFrameworkCore.InMemory")
+                    {
+                        context.Database.EnsureDeleted();
+                        context.Database.EnsureCreated();
+                    }
+                    else
+                    {
+                        context.Database.Migrate();
+                    }
                 }
             }
